requirements.txt
google-generativeai>=0.8.3
yfinance
pytz
requests
beautifulsoup4